# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "AggregateId": {
        "description": "Stable aggregate identifier. A UUID when the aggregate has no natural key, the natural key when it has one (a Measurement is identified by its local date).",
        "examples": [
          "2026-08-25"
        ],
        "maxLength": 64,
        "minLength": 1,
        "pattern": "^[A-Za-z0-9._:-]+$",
        "type": "string"
      },
      "AggregateMeta": {
        "description": "Synchronisation metadata carried by every aggregate (PRD section 12.1).",
        "properties": {
          "createdAt": {
            "$ref": "#/components/schemas/Instant"
          },
          "deletedAt": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Instant"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "$ref": "#/components/schemas/AggregateId"
          },
          "lastMutationId": {
            "$ref": "#/components/schemas/MutationId"
          },
          "originId": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "originType": {
            "$ref": "#/components/schemas/OriginType"
          },
          "revision": {
            "$ref": "#/components/schemas/Revision"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/Instant"
          }
        },
        "required": [
          "createdAt",
          "deletedAt",
          "id",
          "lastMutationId",
          "originId",
          "originType",
          "revision",
          "updatedAt"
        ],
        "type": "object"
      },
      "AggregateType": {
        "description": "The synchronised aggregate kinds. V1 carries one.",
        "enum": [
          "measurement"
        ],
        "type": "string"
      },
      "CounterString": {
        "description": "A 64-bit counter in canonical decimal form, carried as a string.",
        "examples": [
          "0",
          "1",
          "18446744073709551615"
        ],
        "maxLength": 20,
        "pattern": "^(?:0|[1-9]\\d*)$",
        "type": "string"
      },
      "Cursor": {
        "description": "Opaque pagination cursor. Store and return it verbatim; never parse it.",
        "examples": [
          "eyJ2IjoxLCJzZXEiOiI0MiJ9"
        ],
        "maxLength": 512,
        "minLength": 1,
        "pattern": "^[A-Za-z0-9_-]+$",
        "type": "string"
      },
      "CursorPayload": {
        "description": "Decoded cursor. Server-side only: clients receive and return the encoded string.",
        "properties": {
          "seq": {
            "$ref": "#/components/schemas/Sequence"
          },
          "v": {
            "const": 1,
            "type": "number"
          }
        },
        "required": [
          "seq",
          "v"
        ],
        "type": "object"
      },
      "DeleteChange": {
        "properties": {
          "aggregateId": {
            "$ref": "#/components/schemas/AggregateId"
          },
          "aggregateType": {
            "$ref": "#/components/schemas/AggregateType"
          },
          "meta": {
            "$ref": "#/components/schemas/AggregateMeta"
          },
          "op": {
            "const": "delete",
            "type": "string"
          },
          "payload": {
            "type": "null"
          },
          "payloadSchemaVersion": {
            "$ref": "#/components/schemas/PayloadSchemaVersion"
          },
          "sequence": {
            "$ref": "#/components/schemas/Sequence"
          }
        },
        "required": [
          "aggregateId",
          "aggregateType",
          "meta",
          "op",
          "payload",
          "payloadSchemaVersion",
          "sequence"
        ],
        "type": "object"
      },
      "DeleteMutation": {
        "description": "Deletion of any aggregate. Produces a tombstone, never a hard erase.",
        "properties": {
          "aggregateId": {
            "$ref": "#/components/schemas/AggregateId"
          },
          "aggregateType": {
            "$ref": "#/components/schemas/AggregateType"
          },
          "baseRevision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Revision"
              },
              {
                "type": "null"
              }
            ]
          },
          "clientOccurredAt": {
            "$ref": "#/components/schemas/Instant"
          },
          "mutationId": {
            "$ref": "#/components/schemas/MutationId"
          },
          "op": {
            "const": "delete",
            "type": "string"
          },
          "origin": {
            "$ref": "#/components/schemas/Origin"
          },
          "payload": {
            "type": "null"
          },
          "payloadSchemaVersion": {
            "$ref": "#/components/schemas/PayloadSchemaVersion"
          }
        },
        "required": [
          "aggregateId",
          "aggregateType",
          "baseRevision",
          "clientOccurredAt",
          "mutationId",
          "op",
          "origin",
          "payload",
          "payloadSchemaVersion"
        ],
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/MueError"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
      "Instant": {
        "description": "ISO-8601 instant in UTC, always suffixed with Z.",
        "examples": [
          "2026-08-25T09:41:00.000Z"
        ],
        "format": "date-time",
        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
        "type": "string"
      },
      "LivenessReport": {
        "properties": {
          "status": {
            "const": "ok",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      },
      "LocalDate": {
        "description": "ISO-8601 calendar date, no time and no zone.",
        "examples": [
          "2026-08-25"
        ],
        "format": "date",
        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$",
        "type": "string"
      },
      "MeasurementPayloadV1": {
        "description": "Weight measurement, payload schema version 1.",
        "properties": {
          "date": {
            "$ref": "#/components/schemas/LocalDate"
          },
          "weightCg": {
            "maximum": 25000,
            "minimum": 3000,
            "multipleOf": 5,
            "type": "integer"
          }
        },
        "required": [
          "date",
          "weightCg"
        ],
        "type": "object"
      },
      "MeasurementUpsertChange": {
        "properties": {
          "aggregateId": {
            "$ref": "#/components/schemas/LocalDate"
          },
          "aggregateType": {
            "const": "measurement",
            "type": "string"
          },
          "meta": {
            "$ref": "#/components/schemas/AggregateMeta"
          },
          "op": {
            "const": "upsert",
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/MeasurementPayloadV1"
          },
          "payloadSchemaVersion": {
            "const": 1,
            "type": "number"
          },
          "sequence": {
            "$ref": "#/components/schemas/Sequence"
          }
        },
        "required": [
          "aggregateId",
          "aggregateType",
          "meta",
          "op",
          "payload",
          "payloadSchemaVersion",
          "sequence"
        ],
        "type": "object"
      },
      "MeasurementUpsertMutation": {
        "description": "Upsert of a weight measurement, payload schema version 1.",
        "properties": {
          "aggregateId": {
            "$ref": "#/components/schemas/LocalDate"
          },
          "aggregateType": {
            "const": "measurement",
            "type": "string"
          },
          "baseRevision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Revision"
              },
              {
                "type": "null"
              }
            ]
          },
          "clientOccurredAt": {
            "$ref": "#/components/schemas/Instant"
          },
          "mutationId": {
            "$ref": "#/components/schemas/MutationId"
          },
          "op": {
            "const": "upsert",
            "type": "string"
          },
          "origin": {
            "$ref": "#/components/schemas/Origin"
          },
          "payload": {
            "$ref": "#/components/schemas/MeasurementPayloadV1"
          },
          "payloadSchemaVersion": {
            "const": 1,
            "type": "number"
          }
        },
        "required": [
          "aggregateId",
          "aggregateType",
          "baseRevision",
          "clientOccurredAt",
          "mutationId",
          "op",
          "origin",
          "payload",
          "payloadSchemaVersion"
        ],
        "type": "object"
      },
      "MueError": {
        "description": "Structured, actionable business error (FR-SYNC-007).",
        "properties": {
          "aggregateId": {
            "$ref": "#/components/schemas/AggregateId"
          },
          "aggregateType": {
            "$ref": "#/components/schemas/AggregateType"
          },
          "code": {
            "$ref": "#/components/schemas/MueErrorCode"
          },
          "currentRevision": {
            "$ref": "#/components/schemas/Revision"
          },
          "field": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "message": {
            "maxLength": 2000,
            "minLength": 1,
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "type": "object"
      },
      "MueErrorCode": {
        "description": "Stable machine-readable code. Known values: auth.forbidden, auth.unauthenticated, http.not_found, server.internal, server.unavailable, sync.aggregate_deleted, sync.invalid_cursor, sync.invalid_payload, sync.missing_required_field, sync.revision_conflict, sync.unknown_aggregate_type, sync.upgrade_required. Treat an unknown code as a non-retryable failure.",
        "examples": [
          "auth.forbidden",
          "auth.unauthenticated",
          "http.not_found",
          "server.internal",
          "server.unavailable",
          "sync.aggregate_deleted",
          "sync.invalid_cursor",
          "sync.invalid_payload",
          "sync.missing_required_field",
          "sync.revision_conflict",
          "sync.unknown_aggregate_type",
          "sync.upgrade_required"
        ],
        "maxLength": 100,
        "pattern": "^[a-z][a-z0-9]*(?:\\.[a-z][a-z0-9_]*)+$",
        "type": "string"
      },
      "MutationApplied": {
        "properties": {
          "mutationId": {
            "$ref": "#/components/schemas/MutationId"
          },
          "revision": {
            "$ref": "#/components/schemas/Revision"
          },
          "sequence": {
            "$ref": "#/components/schemas/Sequence"
          },
          "status": {
            "const": "applied",
            "type": "string"
          }
        },
        "required": [
          "mutationId",
          "revision",
          "sequence",
          "status"
        ],
        "type": "object"
      },
      "MutationDuplicate": {
        "properties": {
          "mutationId": {
            "$ref": "#/components/schemas/MutationId"
          },
          "revision": {
            "$ref": "#/components/schemas/Revision"
          },
          "sequence": {
            "$ref": "#/components/schemas/Sequence"
          },
          "status": {
            "const": "duplicate",
            "type": "string"
          }
        },
        "required": [
          "mutationId",
          "revision",
          "sequence",
          "status"
        ],
        "type": "object"
      },
      "MutationEnvelope": {
        "description": "One mutation (PRD section 12.2). Replaying it never repeats its effect.",
        "oneOf": [
          {
            "$ref": "#/components/schemas/MeasurementUpsertMutation"
          },
          {
            "$ref": "#/components/schemas/DeleteMutation"
          }
        ]
      },
      "MutationId": {
        "description": "UUIDv7 identifying one mutation, globally unique and creation-ordered.",
        "examples": [
          "0198f0a1-2b3c-7d4e-8f90-a1b2c3d4e5f6"
        ],
        "format": "uuid",
        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-7[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12})$",
        "type": "string"
      },
      "MutationOp": {
        "description": "An upsert carries the full aggregate; a delete carries no payload.",
        "enum": [
          "delete",
          "upsert"
        ],
        "type": "string"
      },
      "MutationRejected": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/MueError"
          },
          "mutationId": {
            "$ref": "#/components/schemas/MutationId"
          },
          "status": {
            "const": "rejected",
            "type": "string"
          }
        },
        "required": [
          "error",
          "mutationId",
          "status"
        ],
        "type": "object"
      },
      "MutationResult": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/MutationApplied"
          },
          {
            "$ref": "#/components/schemas/MutationDuplicate"
          },
          {
            "$ref": "#/components/schemas/MutationRejected"
          }
        ]
      },
      "Origin": {
        "description": "The identity that authored a mutation, kept for audit (PRD section 14.7).",
        "properties": {
          "id": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/OriginType"
          }
        },
        "required": [
          "id",
          "type"
        ],
        "type": "object"
      },
      "OriginType": {
        "description": "Who authored a change (PRD section 12.1).",
        "enum": [
          "agent",
          "android",
          "server"
        ],
        "type": "string"
      },
      "PayloadSchemaVersion": {
        "description": "Version of the payload schema for one aggregate type.",
        "examples": [
          1
        ],
        "maximum": 1000,
        "minimum": 1,
        "type": "integer"
      },
      "PullPage": {
        "description": "A page of changes.",
        "properties": {
          "changes": {
            "items": {
              "$ref": "#/components/schemas/SyncChange"
            },
            "type": "array"
          },
          "hasMore": {
            "type": "boolean"
          },
          "lastAndroidSyncAt": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Instant"
              },
              {
                "type": "null"
              }
            ]
          },
          "nextCursor": {
            "$ref": "#/components/schemas/Cursor"
          },
          "serverTime": {
            "$ref": "#/components/schemas/Instant"
          },
          "status": {
            "const": "ok",
            "type": "string"
          }
        },
        "required": [
          "changes",
          "hasMore",
          "lastAndroidSyncAt",
          "nextCursor",
          "serverTime",
          "status"
        ],
        "type": "object"
      },
      "PullRequest": {
        "description": "Reads the journal after `cursor`. `supportedSchemaVersions` is required: it is what turns PRD section 12.4 into a server-enforced invariant.",
        "properties": {
          "cursor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Cursor"
              },
              {
                "type": "null"
              }
            ]
          },
          "limit": {
            "maximum": 500,
            "minimum": 1,
            "type": "integer"
          },
          "supportedSchemaVersions": {
            "$ref": "#/components/schemas/SupportedSchemaVersions"
          }
        },
        "required": [
          "cursor",
          "supportedSchemaVersions"
        ],
        "type": "object"
      },
      "PullResponse": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/PullPage"
          },
          {
            "$ref": "#/components/schemas/PullUpgradeRequired"
          }
        ]
      },
      "PullUpgradeRequired": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/MueError"
          },
          "lastAndroidSyncAt": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Instant"
              },
              {
                "type": "null"
              }
            ]
          },
          "serverTime": {
            "$ref": "#/components/schemas/Instant"
          },
          "status": {
            "const": "upgrade_required",
            "type": "string"
          }
        },
        "required": [
          "error",
          "lastAndroidSyncAt",
          "serverTime",
          "status"
        ],
        "type": "object"
      },
      "PushRequest": {
        "description": "A batch of local mutations. Order within the batch is the outbox order.",
        "properties": {
          "mutations": {
            "items": {
              "$ref": "#/components/schemas/MutationEnvelope"
            },
            "maxItems": 200,
            "minItems": 1,
            "type": "array"
          }
        },
        "required": [
          "mutations"
        ],
        "type": "object"
      },
      "PushResponse": {
        "description": "Exactly one result per submitted mutation, in submission order.",
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/MutationResult"
            },
            "type": "array"
          },
          "serverTime": {
            "$ref": "#/components/schemas/Instant"
          }
        },
        "required": [
          "results",
          "serverTime"
        ],
        "type": "object"
      },
      "ReadinessCheckReport": {
        "properties": {
          "name": {
            "maxLength": 64,
            "minLength": 1,
            "type": "string"
          },
          "status": {
            "enum": [
              "failed",
              "ok"
            ],
            "type": "string"
          }
        },
        "required": [
          "name",
          "status"
        ],
        "type": "object"
      },
      "ReadinessReport": {
        "properties": {
          "checks": {
            "items": {
              "$ref": "#/components/schemas/ReadinessCheckReport"
            },
            "type": "array"
          },
          "status": {
            "enum": [
              "not_ready",
              "ready"
            ],
            "type": "string"
          }
        },
        "required": [
          "checks",
          "status"
        ],
        "type": "object"
      },
      "Revision": {
        "$ref": "#/components/schemas/CounterString",
        "description": "Per-aggregate revision. Compared for optimistic concurrency; never used as a cursor.",
        "examples": [
          "1",
          "42"
        ]
      },
      "Sequence": {
        "$ref": "#/components/schemas/CounterString",
        "description": "Global per-user journal position. Strictly increasing. Ordering only; never compared to a revision.",
        "examples": [
          "1",
          "9007199254740993"
        ]
      },
      "SupportedSchemaVersions": {
        "additionalProperties": {
          "items": {
            "$ref": "#/components/schemas/PayloadSchemaVersion"
          },
          "maxItems": 32,
          "minItems": 1,
          "type": "array"
        },
        "description": "Aggregate type to the payload schema versions the client can apply.",
        "examples": [
          {
            "measurement": [
              1
            ]
          }
        ],
        "propertyNames": {
          "maxLength": 64,
          "minLength": 1,
          "type": "string"
        },
        "type": "object"
      },
      "SyncChange": {
        "description": "One accepted change from the server journal (PRD section 12.3).",
        "oneOf": [
          {
            "$ref": "#/components/schemas/MeasurementUpsertChange"
          },
          {
            "$ref": "#/components/schemas/DeleteChange"
          }
        ]
      }
    },
    "securitySchemes": {
      "androidBearer": {
        "description": "Better Auth bearer token, one session per device so revocation is per-device (PRD section 15.3).",
        "scheme": "bearer",
        "type": "http"
      },
      "webSession": {
        "description": "HttpOnly, Secure, SameSite session cookie, checked server-side.",
        "in": "cookie",
        "name": "better-auth.session_token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "The `/api/v1` contract the Mue Android client is written against, plus the operational health checks. TanStack Start server functions and MCP tools are deliberately outside it (PRD section 20.4).",
    "title": "Mue Platform API",
    "version": "1.0.0"
  },
  "openapi": "3.1.1",
  "paths": {
    "/api/v1/sync/pull": {
      "post": {
        "description": "Returns a page of changes, or `upgrade_required` with no cursor at all when the server holds a payload version the caller did not declare (PRD sections 12.4 and 18).",
        "operationId": "syncPull",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PullRequest"
              }
            }
          },
          "description": "Cursor and capabilities.",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PullResponse"
                }
              }
            },
            "description": "A page of changes, or an upgrade demand."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The cursor is unreadable."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid or revoked credentials."
          }
        },
        "summary": "Read the change journal after a cursor.",
        "tags": [
          "sync"
        ]
      }
    },
    "/api/v1/sync/push": {
      "post": {
        "description": "Applies each mutation independently. A rejected mutation never blocks the rest (FR-SYNC-007), and replaying an already-applied `mutationId` returns the stored result (FR-SYNC-006).",
        "operationId": "syncPush",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushRequest"
              }
            }
          },
          "description": "The outbox batch.",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushResponse"
                }
              }
            },
            "description": "One result per submitted mutation."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The batch itself is malformed."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid or revoked credentials."
          }
        },
        "summary": "Submit local mutations.",
        "tags": [
          "sync"
        ]
      }
    },
    "/health/live": {
      "get": {
        "operationId": "healthLive",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LivenessReport"
                }
              }
            },
            "description": "The process is running."
          }
        },
        "security": [],
        "summary": "Is the process running.",
        "tags": [
          "health"
        ]
      }
    },
    "/health/ready": {
      "get": {
        "operationId": "healthReady",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessReport"
                }
              }
            },
            "description": "Every dependency answered."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessReport"
                }
              }
            },
            "description": "At least one dependency did not."
          }
        },
        "security": [],
        "summary": "Can the process serve traffic.",
        "tags": [
          "health"
        ]
      }
    }
  },
  "security": [
    {
      "androidBearer": []
    },
    {
      "webSession": []
    }
  ],
  "servers": [
    {
      "description": "The private Mue Platform deployment.",
      "url": "/"
    }
  ],
  "tags": [
    {
      "description": "Bidirectional synchronisation (PRD section 11).",
      "name": "sync"
    },
    {
      "description": "Liveness and readiness (PRD section 20.5).",
      "name": "health"
    }
  ]
}